backend/requirements.txt
flask
flask-cors

scikit-learn
joblib
nltk
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
pydantic[email]
python-dotenv
pandas
